ref(profiling): Introduce flamegraph renderer base class
This introduces a base `FlamegraphRenderer` class as the parent class of
`FlamegraphRendererWebGL`, `FlamegraphRenderer2D` and `FlamegraphRendererDOM`.
